refactor(basic-messages): reorganize into protocol/v1 and protocol/v2 subdirs
