+++ packages/didcomm/src/modules/basic-messages/__tests__/BasicMessageService.test.ts
@@ -3,7 +3,8 @@ import { EventEmitter } from '../../../../../core/src/agent/EventEmitter'
 import { getAgentContext, getMockConnection } from '../../../../../core/tests/helpers'
 import { DidCommInboundMessageContext } from '../../../models'
 import { DidCommBasicMessageRole } from '../DidCommBasicMessageRole'
-import { DidCommBasicMessage, DidCommBasicMessageV2 } from '../messages'
+import { DidCommBasicMessage } from '../protocol/v1'
+import { DidCommBasicMessageV2 } from '../protocol/v2'
 import { DidCommBasicMessageRecord } from '../repository/DidCommBasicMessageRecord'
 import { DidCommBasicMessageRepository } from '../repository/DidCommBasicMessageRepository'
 import { DidCommBasicMessageService } from '../services'
