+++ packages/didcomm/src/modules/basic-messages/index.ts
@@ -3,6 +3,6 @@ export * from './DidCommBasicMessageRole'
 export * from './DidCommBasicMessagesApi'
 export * from './DidCommBasicMessagesModule'
 export * from './DidCommBasicMessagesModuleConfig'
-export * from './messages'
+export * from './protocol'
 export * from './repository'
 export * from './services'
