+++ packages/didcomm/src/modules/basic-messages/protocol/index.ts
@@ -1,0 +1,2 @@
+export * from './v1'
+export * from './v2'
