+++ packages/didcomm/src/modules/basic-messages/protocol/v1/handlers/BasicMessageHandler.ts
@@ -1,6 +1,6 @@
-import type { DidCommMessageHandler, DidCommMessageHandlerInboundMessage } from '../../../handlers'
+import type { DidCommMessageHandler, DidCommMessageHandlerInboundMessage } from '../../../../../handlers'
+import type { DidCommBasicMessageService } from '../../../services/DidCommBasicMessageService'
 import { DidCommBasicMessage } from '../messages'
-import type { DidCommBasicMessageService } from '../services/DidCommBasicMessageService'
 
 export class DidCommBasicMessageHandler implements DidCommMessageHandler {
   private basicMessageService: DidCommBasicMessageService
