+++ packages/didcomm/src/modules/basic-messages/protocol/v1/handlers/index.ts
@@ -1,0 +1,1 @@
+export * from './BasicMessageHandler'
