+++ packages/didcomm/src/modules/basic-messages/protocol/v1/index.ts
@@ -1,0 +1,1 @@
+export * from './messages'
