+++ packages/didcomm/src/modules/basic-messages/protocol/v2/handlers/index.ts
@@ -1,2 +1,1 @@
-export * from './BasicMessageHandler'
 export * from './BasicMessageV2Handler'
