+++ packages/didcomm/src/modules/basic-messages/protocol/v2/index.ts
@@ -1,0 +1,1 @@
+export * from './messages'
