+++ packages/didcomm/src/modules/basic-messages/protocol/v2/messages/index.ts
@@ -1,2 +1,1 @@
-export * from './DidCommBasicMessage'
 export * from './DidCommBasicMessageV2'
